+++ ios/InteractiveReader/InteractiveReader/Features/Playback/JobPlaybackView+NowPlaying.swift
@@ -452,6 +452,7 @@ extension JobPlaybackView {
                     #if DEBUG
                     e2eReaderPauseConfirmationCount += 1
                     #endif
+                    clearPendingInteractiveAutoplay(reason: "\(source)PauseConfirmed")
                     playbackTransportDebugLog(
                         "[PlaybackTransport] Job confirmed reader pause source=\(source) requested=false playing=false musicPlaying=\(musicOwnership.isPlaying) systemMusicPlaying=\(musicOwnership.isSystemPlaybackPlaying)"
                     )